+++ contracts/samples/TestAggregatedWallet.sol
@@ -20,7 +20,7 @@ contract TestAggregatedWallet is SimpleWallet, IAggregatedWallet {
     }
 
     function _validateSignature(UserOperation calldata userOp, bytes32 requestId, address userOpAggregator)
-    internal override view returns (uint deadline) {
+    internal override view returns (uint256 deadline) {
         (userOp, requestId);
         require(userOpAggregator == aggregator, "wrong aggregator");
         return 0;
